add module network, change -fs to flag, add appending accuracies to file, save predictions to new predictions folder
