+++ pythonProject/embedding_classifier.py
@@ -8,17 +8,17 @@ import seaborn as sns
 
 import matplotlib
 import numpy as np
+from ann import mean_score_ann, ANN, Data, train_ann
 from matplotlib import pyplot as plt
 from sklearn import svm
 from sklearn.metrics import accuracy_score
-from sklearn.model_selection import train_test_split, GridSearchCV
+from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score
 from sklearn.neighbors import KNeighborsClassifier
 from sklearn.preprocessing import StandardScaler
 from sklearn.svm import SVC
 from torch import nn
 from torch.utils.data import DataLoader
-from utils import feature_selected_sets, ANN, Data, \
-    train_ann, save_preds, NP_SEED
+from utils import NP_SEED, get_feature_names, all_subsets
 
 matplotlib.use('TkAgg')
 
@@ -137,13 +137,64 @@ class EmbeddingClassifier:
         return accuracy
 
 
+def save_preds(preds, labels, clf, dataset_name, feature_selection):
+    """saves labels and predictions to a csv-file"""
+    data = {"preds": preds, "labels": labels}
+    df = pd.DataFrame(data)
+    df.to_csv(f'predictions/preds_labels_{clf}_{dataset_name}_fs{feature_selection}.csv', index=False)
+
+
+def get_best_feature_set(clf, X_train, y):
+    """returns the best set for classifying using an SVM/KNN clf, uses cross-validation and takes the set with the
+    highest mean accuracy"""
+    n_features = X_train.shape[1]
+    best_score = -np.inf
+    best_subset = None
+    count = 1
+    if isinstance(clf, (KNeighborsClassifier, SVC)):
+        for subset in all_subsets(np.arange(n_features)):
+            score = cross_val_score(clf, np.reshape(X_train[:, subset], (X_train.shape[0], -1)), y, cv=5).mean()
+            if score > best_score:
+                best_score, best_subset = score, subset
+            print(f'subset {count}/{2 ** n_features - 1}')
+            count += 1
+    else:
+        for subset in all_subsets(np.arange(n_features)):
+            score = mean_score_ann(np.reshape(X_train[:, subset], (X_train.shape[0], -1)), y)
+            if score > best_score:
+                best_score, best_subset = score, subset
+            print(f'subset {count}/{2 ** n_features - 1}')
+            count += 1
+
+    print(f"best_subset: {best_subset.numpy()} with best score: {best_score}")
+    return best_subset.numpy(), best_score
+
+
+def feature_selected_sets(clf, X_train, X_test, y):
+    """returns the modified training and test sets after performing feature selection on them"""
+    best_subset, best_score = get_best_feature_set(clf, X_train, y)
+    features = get_feature_names(best_subset)
+    print(f"The optimal features selected for {type(clf).__name__} were: {features}")
+    X_train_fs = X_train[:, best_subset]
+    X_test_fs = X_test[:, best_subset]
+    assert (X_train_fs.shape[0], len(best_subset)) == X_train_fs.shape
+    return X_train_fs, X_test_fs
+
+
+def append_accuracies(dn, clf, fs, acc):
+    with open('accuracies.txt', mode='a') as file:
+        file.write(f'Accuracy for {dn} {clf} fs={fs}: {acc}\n')
+    file.close()
+
+
+
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument('--dn', type=str, help='The name of the dataset to be classified.')
     parser.add_argument('--clf', type=str, default='knn', help='Which classifier model should be used. Choose '
                                                                'between: svm, knn or ann')
-    parser.add_argument('--fs', default=False, help='Whether the feature selection is wished. Default is true')
-
+    parser.add_argument('--fs', action="store_true", default=False, help='Whether the feature selection is wished. '
+                                                                         'The default is False')
     args = parser.parse_args()
     if args.dn is None:
         raise argparse.ArgumentError(None, "Please enter the required arguments: --dn, --clf and optionally --fs")
@@ -156,12 +207,15 @@ if __name__ == "__main__":
     if clf_model.lower() == 'knn':
         acc = embedding_classifier.predict_knn()
         print(f"Accuracy for our testing {dataset_name} dataset with tuning using the KNN model is: {acc}")
+        append_accuracies(dataset_name, clf_model, feature_selection, acc)
     elif clf_model.lower() == 'svm':
         acc = embedding_classifier.predict_svm()
         print(f"Accuracy for our testing {dataset_name} dataset with tuning using the SVM model is: {acc}")
+        append_accuracies(dataset_name, clf_model, feature_selection, acc)
     elif clf_model.lower() == 'ann':
         acc = embedding_classifier.predict_ann()
         print(f"Accuracy for our testing {dataset_name} dataset with tuning using the ANN model is: {acc}")
+        append_accuracies(dataset_name, clf_model, feature_selection, acc)
     else:
         raise argparse.ArgumentTypeError('Invalid classifier. Pick between knn, svm or ann.')
